\name{FMCCSD-package}
\alias{FMCCSD-package}
\alias{FMCCSD}
\docType{package}
\title{
\packageTitle{FMCCSD}
}
\description{
%\packageDescription{CRFMCCS}
            Current status data abounds
            in the field of epidemiology and public health, where the only observable data for a subject is
            the random inspection time and the event status at inspection. Motivated by such a current status
            data from a periodontal study where data are inherently clustered,
            we propose a unified methodology to analyze such complex data.
}

\author{
\packageAuthor{FMCCSD}

Maintainer:\packageMaintainer{FMCCSD}
}

%\references{
%Wang, T., He, K., Ma, W., Bandyopadhyay, D. and Sinha S. Mixture Cure frailty Generalized Odds Ratio Model for Clustered Current Status Data. Preprint. 
%}
