\name{PD}
\alias{PD}
\title{
A clustered current status dataset.
}
\description{
A clustered current status dataset arises from a periodontal disease (PD) study where tooth level data are clusterd within subjects. The first and second columns are the index for patients and the inspection time for each tooth, respectively. The next 3 column are the three subject level covariates (gender, smoking and Hba1c). After that, it is the tooth level covariate (jaw). The last column is the indicator for the event.
}
\usage{
data(PD)
}

\seealso{
\code{\link{CSDfit}}
}
